+++ src/frontend/hooks/MongoDB.jsx
@@ -1,0 +1,1 @@
+export const [collectionData, setCollectionData] = React.useState([]);
