+++ src/frontend/main.jsx
@@ -1,6 +1,7 @@
 import React from 'react'
 import ReactDOM from 'react-dom/client'
 import App from './App.jsx'
+import Login from "./front-end/screens/Login.js"
 import './css/index.css'
 
 ReactDOM.createRoot(document.getElementById('root')).render(
